requirements.txt
Django>=5.0,<6.0
django-environ>=0.11.2
django-crispy-forms>=2.1
crispy-bootstrap5>=0.7
django-jazzmin>=2.6.0
django-import-export>=3.3.1
Pillow>=10.0.0
psycopg2-binary>=2.9.9
whitenoise>=6.5.0
gunicorn>=21.2.0
dj-database-url>=2.1.0
python-dotenv>=1.0.0
cloudinary>=1.36.0
django-cloudinary-storage>=0.3.0
